requirements.txt
streamlit
biopython
python-docx
scipy
numpy
pandas
langchain
langchain-community